app-health-back/src/@api/app-health/application/resolvers/app-health-paginate-applications.resolver.ts
import { AppHealthPaginateApplicationsHandler } from '@api/app-health/application';
import { Pagination } from '@api/graphql';
import { QueryStatement, Timezone } from '@aurorajs.dev/core';
import { Args, Query, Resolver } from '@nestjs/graphql';

@Resolver()
export class AppHealthPaginateApplicationsResolver
{
    constructor(
        private readonly handler: AppHealthPaginateApplicationsHandler,
    ) {}

    @Query('appHealthPaginateApplications')
    async main(
        @Args('query') queryStatement?: QueryStatement,
        @Args('constraint') constraint?: QueryStatement,
        @Timezone() timezone?: string,
    ): Promise<Pagination>
    {
        return await this.handler.main(
            queryStatement,
            constraint,
            timezone,
        );
    }
}
